Add emailsUtilityService and emailByPurposeService

+++ backend/src/main/java/co/oleh/realperfect/objectreview/ObjectReviewService.java
@@ -1,13 +1,14 @@
 package co.oleh.realperfect.objectreview;
 
 import co.oleh.realperfect.auth.SpringSecurityUser;
-import co.oleh.realperfect.emails.EmailSenderService;
+import co.oleh.realperfect.emails.EmailsByPurposeService;
 import co.oleh.realperfect.mapping.MyObjectReviewDto;
 import co.oleh.realperfect.mapping.ObjectReviewDto;
 import co.oleh.realperfect.mapping.mappers.MappingService;
 import co.oleh.realperfect.model.ObjectReview;
 import co.oleh.realperfect.model.Realtor;
 import co.oleh.realperfect.model.RealtyObject;
+import co.oleh.realperfect.model.user.RoleUtils;
 import co.oleh.realperfect.model.user.User;
 import co.oleh.realperfect.repository.ObjectReviewRepository;
 import co.oleh.realperfect.repository.RealtyObjectCrudRepository;
@@ -38,7 +39,7 @@ public class ObjectReviewService {
 
     private final RealtyObjectCrudRepository realtyObjectRepository;
 
-    private EmailSenderService emailService;
+    private EmailsByPurposeService emailService;
     private UserRepository userRepository;
     private ObjectReviewRepository objectReviewRepository;
     private MappingService mappingService;
@@ -86,7 +87,8 @@ public class ObjectReviewService {
 
             objectReviewRepository.deleteById(objectReview.getId());
 
-            emailService.sendObjectReviewCancelAsync("Reviews Removed For RealtyObject", userFromDb, objectReview, realtyObject, realtor);
+            emailService.sendObjectReviewCancelAsync("Reviews Removed For RealtyObject", userFromDb, objectReview,
+                    realtyObject, realtor);
         }
 
         return objectReviews;
@@ -153,11 +155,7 @@ public class ObjectReviewService {
         Instant currentHourToCheck = houseOpeningTime;
         Instant twoHoursFromNow = Instant.now().plus(2, ChronoUnit.HOURS);
         while (houseClosingTime.isAfter(currentHourToCheck)) {
-            if (currentHourToCheck.isBefore(twoHoursFromNow)) {
-                break;
-            }
-            if (!busyTimes.contains(currentHourToCheck) &&
-                    currentHourToCheck.isAfter(Instant.now())) {
+            if (currentHourToCheck.isAfter(twoHoursFromNow) && !busyTimes.contains(currentHourToCheck)) {
                 availableSlots.add(currentHourToCheck);
             }
             currentHourToCheck = currentHourToCheck.plus(1, ChronoUnit.HOURS);
@@ -167,17 +165,28 @@ public class ObjectReviewService {
     }
 
     @Transactional
-    public int approveReviewById(Long objectReviewId) {
-        // send email
-        return objectReviewRepository.updateApprovedStatus(objectReviewId, true);
+    public int approveReviewById(Long objectReviewId, SpringSecurityUser user) {
+        int approveResult = objectReviewRepository.updateApprovedStatus(objectReviewId, true);
+        User userFromDb = userRepository.findById(user.getId()).get();
+
+        ObjectReview objectReview = this.objectReviewRepository.findById(objectReviewId).get();
+        this.emailService.sendObjectReviewApprovedAsync(userFromDb, objectReview, objectReview.getRealtyObj(),
+                objectReview.getRealtor());
+
+        return approveResult;
     }
 
     @Transactional
     public void deleteReviewById(Long reviewId, SpringSecurityUser user, String reason) {
-        // TODO verify user
         User userFromDb = userRepository.findById(user.getId()).get();
         ObjectReview objectReview = this.objectReviewRepository.findById(reviewId).get();
-        this.emailService.sendObjectReviewCancelAsync(reason, userFromDb, objectReview, objectReview.getRealtyObj(), objectReview.getRealtor());
-        this.objectReviewRepository.deleteById(reviewId);
+        if (RoleUtils.containsAuthority(user, RoleUtils.REALTOR_ROLE) ||
+                objectReview.getRealtyObj().getOwner().getId().equals(user.getId())) {
+            this.objectReviewRepository.deleteById(reviewId);
+            this.emailService.sendObjectReviewCancelAsync(reason, userFromDb, objectReview, objectReview.getRealtyObj(),
+                    objectReview.getRealtor());
+        } else {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Insufficient permissions");
+        }
     }
 }
